Extract ProblemExceptionInterface implementation tests to abstract test case
